# backend-go/docs/swagger.yaml
basePath: /api
definitions:
...
  features_payments_presentation.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  presentation.CreateBookingPaymentRequest:
    properties:
      amount_cents:
        type: integer
      booking_id:
        type: integer
      customer_id:
        type: string
      user_id:
        type: integer
    required:
    - amount_cents
    - booking_id
    - customer_id
    - user_id
    type: object
...
  presentation.CreateClassPaymentRequest:
    properties:
      amount_cents:
        type: integer
      customer_id:
        type: string
      enrollment_id:
        type: integer
      user_id:
        type: integer
    required:
    - amount_cents
    - customer_id
    - enrollment_id
    - user_id
    type: object
...
  presentation.CreateClubPaymentRequest:
    properties:
      amount_cents:
        type: integer
      customer_id:
        type: string
      membership_id:
        type: integer
      user_id:
        type: integer
    required:
    - amount_cents
    - customer_id
    - membership_id
    - user_id
    type: object
...
  presentation.CreatePaymentRequest:
    properties:
      amount_cents:
        type: integer
      customer_id:
        type: string
      description:
        type: string
      user_id:
        type: integer
    required:
    - amount_cents
    - customer_id
    - description
    - user_id
    type: object
  presentation.CreatePistaRequest:
    properties:
      image_url:
        type: string
      nombre:
        maxLength: 100
        minLength: 1
        type: string
      precio_hora_base:
        minimum: 0
        type: number
      superficie:
        type: string
      tipo:
        enum:
        - FUTBOL
        - TENIS
        - BALONCESTO
        - PADEL
        - POLIDEPORTIVA
        type: string
    required:
    - nombre
    - precio_hora_base
    - tipo
    type: object
...
  presentation.PaymentResponse:
    properties:
      amount_cents:
        type: integer
      amount_euros:
        type: number
      booking_id:
        type: integer
      class_enrollment_id:
        type: integer
      club_membership_id:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      payment_type:
        type: string
      provider:
        type: string
      status:
        type: string
      stripe_payment_intent_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  presentation.PistaPagedResponse:
    properties:
      items:
        description: Tus DTOs de pista normales
        items:
          $ref: '#/definitions/presentation.PistaResponse'
        type: array
      limit:
        description: Items por página
        type: integer
      page:
        description: Página actual
        type: integer
      total:
        description: Total de resultados en BD
        type: integer
      total_pages:
        description: Total de páginas
        type: integer
    type: object
  presentation.PistaResponse:
    properties:
      esActiva:
        type: boolean
      estado:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      nombre:
        type: string
      precioHoraBase:
        type: number
      superficie:
        type: string
      tipo:
        type: string
    type: object
  presentation.RefundPaymentRequest:
    properties:
      payment_id:
        type: integer
    required:
    - payment_id
    type: object
...
  presentation.UpdatePistaRequest:
    properties:
      es_activa:
        type: boolean
      estado:
        enum:
        - DISPONIBLE
        - MANTENIMIENTO
        - INACTIVA
        type: string
      image_url:
        type: string
      nombre:
        maxLength: 100
        minLength: 1
        type: string
      precio_hora_base:
        minimum: 0
        type: number
      superficie:
        type: string
      tipo:
        enum:
        - FUTBOL
        - TENIS
        - BALONCESTO
        - PADEL
        - POLIDEPORTIVA
        type: string
    required:
    - estado
    - nombre
    - precio_hora_base
    - tipo
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API REST para gestión de instalaciones deportivas
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: PoliManage API
  version: "1.0"
paths:
...
  /api/payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del pago
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/presentation.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.PaymentResponse'
      summary: Procesar un pago
      tags:
      - payments
  /api/payments/{id}:
    get:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.PaymentResponse'
      summary: Obtener pago por ID
      tags:
      - payments
  /api/payments/booking:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del pago de reserva
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateBookingPaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.PaymentResponse'
      summary: Procesar pago de reserva
      tags:
      - payments
  /api/payments/class:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del pago de clase
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateClassPaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.PaymentResponse'
      summary: Procesar pago de clase
      tags:
      - payments
  /api/payments/club:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del pago de membresía
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/presentation.CreateClubPaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.PaymentResponse'
      summary: Procesar pago de membresía
      tags:
      - payments
  /api/payments/refund:
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del reembolso
        in: body
        name: refund
        required: true
        schema:
          $ref: '#/definitions/presentation.RefundPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/features_payments_presentation.MessageResponse'
      summary: Reembolsar un pago
      tags:
      - payments
  /api/payments/user/{user_id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.PaymentResponse'
            type: array
      summary: Obtener pagos de usuario
      tags:
      - payments
  /api/pistas:
    get:
      parameters:
      - description: Búsqueda por nombre o ubicación
        in: query
        name: q
        type: string
      - description: Filtro por tipo de deporte (FUTBOL, TENIS, BALONCESTO, PADEL,
          POLIDEPORTIVA)
        in: query
        name: deporte
        type: string
      - description: Precio mínimo por hora (en euros)
        in: query
        name: min_price
        type: integer
      - description: Precio máximo por hora (en euros)
        in: query
        name: max_price
        type: integer
      - default: id_desc
        description: 'Ordenación: precio_asc, precio_desc, nombre_asc, nombre_desc'
        in: query
        name: sort
        type: string
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 12
        description: Items por página
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.PistaPagedResponse'
      summary: Listar pistas con filtros avanzados
      tags:
      - pistas
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos de la pista
        in: body
        name: pista
        required: true
        schema:
          $ref: '#/definitions/presentation.CreatePistaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presentation.PistaResponse'
      summary: Crear una nueva pista
      tags:
      - pistas
  /api/pistas/{id}:
    delete:
      parameters:
      - description: ID de la pista
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Eliminar una pista
      tags:
      - pistas
    get:
      parameters:
      - description: ID de la pista
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.PistaResponse'
      summary: Obtener pista por ID
      tags:
      - pistas
    put:
      consumes:
      - application/json
      parameters:
      - description: ID de la pista
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a actualizar
        in: body
        name: pista
        required: true
        schema:
          $ref: '#/definitions/presentation.UpdatePistaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presentation.PistaResponse'
      summary: Actualizar una pista
      tags:
      - pistas
...
schemes:
- http
swagger: "2.0"